Normalize Res alias and export it and the Error type
